requirements.txt
streamlit
pypdf
chromadb
google-generativeai